Create LICENSE
I've added a license for the code, with the licensees being the four of us working on the repos, plus the CWI. This arrangement is based on what I understood from Peter is the custom, or at least the appropriate fashion, to publish FOSS by CWI employees that is not previously intended to go under another license (e.g. the MonetDB project).